fix #1 upravene aj pre longy

+++ src/main/java/com/mycompany/calcularot/Calculator.java
@@ -2,7 +2,7 @@ package com.mycompany.calcularot;
 
 public class Calculator {
 
-    public long scitaj(int a, int b) {
+    public long scitaj(long a, long b) {
         return a + b;
     }
 }
